requirements.txt
Django==1.11.7
django-rest-framework==0.1.0
djangorestframework==3.7.1
pytz==2017.3
